functions/requirements.txt
functions-framework
google-cloud-firestore
pandas
pytz
statsmodels
firebase-admin
python-dotenv